# Fix/590 docker mongo rs init