--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -6,6 +6,10 @@ All notable changes to the kytos project will be documented in this file.
 UNRELEASED - Under development
 ******************************
 
+Fixed
+=====
+- Enhanced docker mongo setu script (rs-init.sh) to properly run on PRIMARY node
+
 [2025.1.0] - 2025-04-15
 ***********************