Closes #590

### Summary

See updated changelog file and/or add any other summarized helpful information for reviewers

### Local Tests

Before this PR: please refer to the issue for the output before the PR

After the PR:
```
# docker compose logs mongo-setup -t -f
WARN[0000] /var/home-amlight-bkp/vnoviflow-tests/kytos/docker-compose.yml: `version` is obsolete
mongo-rs-init  | 2025-06-10T17:56:39.057459318Z Waiting for mongo nodes serverStatus...
mongo-rs-init  | 2025-06-10T17:56:39.064772869Z All mongo nodes are up
mongo-rs-init  | 2025-06-10T17:56:39.065926377Z Applying replicaSet rs0 config on mongo1:27017 at 17:56:39...
mongo-rs-init  | 2025-06-10T17:56:39.264108628Z Current Mongosh Log ID:	684871d751a71ac43169e327
mongo-rs-init  | 2025-06-10T17:56:39.277625324Z Connecting to:		mongodb://mongo1:27017/?directConnection=true&appName=mongosh+2.5.2
mongo-rs-init  | 2025-06-10T17:56:39.404495608Z MongoNetworkError: connect ECONNREFUSED 172.22.0.4:27017
mongo-rs-init  | 2025-06-10T17:56:39.419293433Z Creating DBs and users on mongo1:27017 at 17:56:39...
mongo-rs-init  | 2025-06-10T17:56:39.606054010Z Current Mongosh Log ID:	684871d72c50cce4b269e327
mongo-rs-init  | 2025-06-10T17:56:39.620132455Z Connecting to:		mongodb://mongo1:27017/?directConnection=true&appName=mongosh+2.5.2
mongo-rs-init  | 2025-06-10T17:56:39.739482554Z MongoNetworkError: connect ECONNREFUSED 172.22.0.4:27017
mongo-rs-init  | 2025-06-10T17:56:41.370154341Z Waiting for mongo nodes serverStatus...
mongo-rs-init  | 2025-06-10T17:56:41.377106997Z All mongo nodes are up
mongo-rs-init  | 2025-06-10T17:56:41.378880694Z Applying replicaSet rs0 config on mongo1:27017 at 17:56:41...
mongo-rs-init  | 2025-06-10T17:56:41.567878635Z Current Mongosh Log ID:	684871d9d56256c26769e327
mongo-rs-init  | 2025-06-10T17:56:41.582123042Z Connecting to:		mongodb://mongo1:27017/?directConnection=true&appName=mongosh+2.5.2
mongo-rs-init  | 2025-06-10T17:56:41.741093325Z Using MongoDB:		7.0.21
mongo-rs-init  | 2025-06-10T17:56:41.741112376Z Using Mongosh:		2.5.2
mongo-rs-init  | 2025-06-10T17:56:41.741127196Z
mongo-rs-init  | 2025-06-10T17:56:41.741133446Z For mongosh info see: https://www.mongodb.com/docs/mongodb-shell/
mongo-rs-init  | 2025-06-10T17:56:41.741138856Z
mongo-rs-init  | 2025-06-10T17:56:41.741143526Z
mongo-rs-init  | 2025-06-10T17:56:41.741148246Z To help improve our products, anonymous usage data is collected and sent to MongoDB periodically (https://www.mongodb.com/legal/privacy-policy).
mongo-rs-init  | 2025-06-10T17:56:41.741152746Z You can opt-out by running the disableTelemetry() command.
mongo-rs-init  | 2025-06-10T17:56:41.741156596Z
mongo-rs-init  | 2025-06-10T17:56:41.779142105Z ------
mongo-rs-init  | 2025-06-10T17:56:41.779158555Z    The server generated these startup warnings when booting
mongo-rs-init  | 2025-06-10T17:56:41.779162915Z    2025-06-10T17:56:39.232+00:00: Access control is not enabled for the database. Read and write access to data and configuration is unrestricted
mongo-rs-init  | 2025-06-10T17:56:41.779166756Z    2025-06-10T17:56:39.232+00:00: You are running this process as the root user, which is not recommended
mongo-rs-init  | 2025-06-10T17:56:41.779170516Z    2025-06-10T17:56:39.234+00:00: You are running on a NUMA machine. We suggest launching mongod like this to avoid performance problems: numactl --interleave=all mongod [other options]
mongo-rs-init  | 2025-06-10T17:56:41.779174276Z    2025-06-10T17:56:39.234+00:00: For customers running MongoDB 7.0, we suggest changing the contents of the following sysfsFile
mongo-rs-init  | 2025-06-10T17:56:41.779177856Z    2025-06-10T17:56:39.234+00:00: vm.max_map_count is too low
mongo-rs-init  | 2025-06-10T17:56:41.779186356Z ------
mongo-rs-init  | 2025-06-10T17:56:41.779200786Z
mongo-rs-init  | 2025-06-10T17:56:42.085809600Z test> ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ... ...
mongo-rs-init  | 2025-06-10T17:56:43.267010785Z test> { ok: 1 }
mongo-rs-init  | 2025-06-10T17:56:43.299681973Z rs0 [direct: other] test> {
mongo-rs-init  | 2025-06-10T17:56:43.299703194Z   set: 'rs0',
mongo-rs-init  | 2025-06-10T17:56:43.299707584Z   date: ISODate('2025-06-10T17:56:43.293Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299711234Z   myState: 2,
mongo-rs-init  | 2025-06-10T17:56:43.299715004Z   term: Long('0'),
mongo-rs-init  | 2025-06-10T17:56:43.299718454Z   syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:43.299721964Z   syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:43.299725394Z   heartbeatIntervalMillis: Long('2000'),
mongo-rs-init  | 2025-06-10T17:56:43.299729024Z   majorityVoteCount: 2,
mongo-rs-init  | 2025-06-10T17:56:43.299732434Z   writeMajorityCount: 2,
mongo-rs-init  | 2025-06-10T17:56:43.299735964Z   votingMembersCount: 3,
mongo-rs-init  | 2025-06-10T17:56:43.299739374Z   writableVotingMembersCount: 3,
mongo-rs-init  | 2025-06-10T17:56:43.299742834Z   optimes: {
mongo-rs-init  | 2025-06-10T17:56:43.299746864Z     lastCommittedOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299750494Z     lastCommittedWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299754084Z     readConcernMajorityOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299757664Z     appliedOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299761214Z     durableOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299764754Z     lastAppliedWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299768425Z     lastDurableWallTime: ISODate('2025-06-10T17:56:42.828Z')
mongo-rs-init  | 2025-06-10T17:56:43.299771955Z   },
mongo-rs-init  | 2025-06-10T17:56:43.299775435Z   lastStableRecoveryTimestamp: Timestamp({ t: 1749578202, i: 1 }),
mongo-rs-init  | 2025-06-10T17:56:43.299778975Z   members: [
mongo-rs-init  | 2025-06-10T17:56:43.299782425Z     {
mongo-rs-init  | 2025-06-10T17:56:43.299785815Z       _id: 1,
mongo-rs-init  | 2025-06-10T17:56:43.299789215Z       name: 'mongo1:27017',
mongo-rs-init  | 2025-06-10T17:56:43.299792635Z       health: 1,
mongo-rs-init  | 2025-06-10T17:56:43.299796035Z       state: 2,
mongo-rs-init  | 2025-06-10T17:56:43.299799485Z       stateStr: 'SECONDARY',
mongo-rs-init  | 2025-06-10T17:56:43.299802915Z       uptime: 6,
mongo-rs-init  | 2025-06-10T17:56:43.299806705Z       optime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299810245Z       optimeDate: ISODate('2025-06-10T17:56:42.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299813745Z       lastAppliedWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299817235Z       lastDurableWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299820865Z       syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:43.299834116Z       syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:43.299838276Z       infoMessage: '',
mongo-rs-init  | 2025-06-10T17:56:43.299841726Z       configVersion: 1,
mongo-rs-init  | 2025-06-10T17:56:43.299845286Z       configTerm: 0,
mongo-rs-init  | 2025-06-10T17:56:43.299848666Z       self: true,
mongo-rs-init  | 2025-06-10T17:56:43.299852146Z       lastHeartbeatMessage: ''
mongo-rs-init  | 2025-06-10T17:56:43.299855586Z     },
mongo-rs-init  | 2025-06-10T17:56:43.299858956Z     {
mongo-rs-init  | 2025-06-10T17:56:43.299862296Z       _id: 2,
mongo-rs-init  | 2025-06-10T17:56:43.299865806Z       name: 'mongo2:27018',
mongo-rs-init  | 2025-06-10T17:56:43.299869186Z       health: 1,
mongo-rs-init  | 2025-06-10T17:56:43.299872636Z       state: 0,
mongo-rs-init  | 2025-06-10T17:56:43.299876016Z       stateStr: 'STARTUP',
mongo-rs-init  | 2025-06-10T17:56:43.299879456Z       uptime: 0,
mongo-rs-init  | 2025-06-10T17:56:43.299882856Z       optime: { ts: Timestamp({ t: 0, i: 0 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299886596Z       optimeDurable: { ts: Timestamp({ t: 0, i: 0 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299890106Z       optimeDate: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299893546Z       optimeDurableDate: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299896987Z       lastAppliedWallTime: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299900447Z       lastDurableWallTime: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299903947Z       lastHeartbeat: ISODate('2025-06-10T17:56:43.079Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299908547Z       lastHeartbeatRecv: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299912127Z       pingMs: Long('0'),
mongo-rs-init  | 2025-06-10T17:56:43.299915537Z       lastHeartbeatMessage: '',
mongo-rs-init  | 2025-06-10T17:56:43.299918957Z       syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:43.299922377Z       syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:43.299925857Z       infoMessage: '',
mongo-rs-init  | 2025-06-10T17:56:43.299929207Z       configVersion: -2,
mongo-rs-init  | 2025-06-10T17:56:43.299932617Z       configTerm: -1
mongo-rs-init  | 2025-06-10T17:56:43.299936047Z     },
mongo-rs-init  | 2025-06-10T17:56:43.299939417Z     {
mongo-rs-init  | 2025-06-10T17:56:43.299942747Z       _id: 3,
mongo-rs-init  | 2025-06-10T17:56:43.299946127Z       name: 'mongo3:27019',
mongo-rs-init  | 2025-06-10T17:56:43.299949627Z       health: 1,
mongo-rs-init  | 2025-06-10T17:56:43.299953007Z       state: 0,
mongo-rs-init  | 2025-06-10T17:56:43.299956357Z       stateStr: 'STARTUP',
mongo-rs-init  | 2025-06-10T17:56:43.299959747Z       uptime: 0,
mongo-rs-init  | 2025-06-10T17:56:43.299963558Z       optime: { ts: Timestamp({ t: 0, i: 0 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299967068Z       optimeDurable: { ts: Timestamp({ t: 0, i: 0 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:43.299970578Z       optimeDate: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299976848Z       optimeDurableDate: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299980768Z       lastAppliedWallTime: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299984268Z       lastDurableWallTime: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299987738Z       lastHeartbeat: ISODate('2025-06-10T17:56:43.079Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299991188Z       lastHeartbeatRecv: ISODate('1970-01-01T00:00:00.000Z'),
mongo-rs-init  | 2025-06-10T17:56:43.299994758Z       pingMs: Long('0'),
mongo-rs-init  | 2025-06-10T17:56:43.299998128Z       lastHeartbeatMessage: '',
mongo-rs-init  | 2025-06-10T17:56:43.300001888Z       syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:43.300005288Z       syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:43.300009348Z       infoMessage: '',
mongo-rs-init  | 2025-06-10T17:56:43.300013648Z       configVersion: -2,
mongo-rs-init  | 2025-06-10T17:56:43.300018408Z       configTerm: -1
mongo-rs-init  | 2025-06-10T17:56:43.300022648Z     }
mongo-rs-init  | 2025-06-10T17:56:43.300027218Z   ],
mongo-rs-init  | 2025-06-10T17:56:43.300031219Z   ok: 1,
mongo-rs-init  | 2025-06-10T17:56:43.300035449Z   '$clusterTime': {
mongo-rs-init  | 2025-06-10T17:56:43.300039699Z     clusterTime: Timestamp({ t: 1749578202, i: 1 }),
mongo-rs-init  | 2025-06-10T17:56:43.300043759Z     signature: {
mongo-rs-init  | 2025-06-10T17:56:43.300048069Z       hash: Binary.createFromBase64('AAAAAAAAAAAAAAAAAAAAAAAAAAA=', 0),
mongo-rs-init  | 2025-06-10T17:56:43.300052899Z       keyId: Long('0')
mongo-rs-init  | 2025-06-10T17:56:43.300057469Z     }
mongo-rs-init  | 2025-06-10T17:56:43.300061899Z   },
mongo-rs-init  | 2025-06-10T17:56:43.300065399Z   operationTime: Timestamp({ t: 1749578202, i: 1 })
mongo-rs-init  | 2025-06-10T17:56:43.300069019Z }
mongo-rs-init  | 2025-06-10T17:56:43.312992486Z rs0 [direct: secondary] test>
mongo-rs-init  | 2025-06-10T17:56:43.331614470Z rs0 [direct: secondary] test>
mongo-rs-init  | 2025-06-10T17:56:43.363593887Z rs0 [direct: secondary] test> ... ... ... ... Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:44.375712865Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:45.381957214Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:46.387737465Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:47.394275178Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:48.400553206Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:49.406529270Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:50.412566615Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:51.418301196Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:52.423886843Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:53.428697639Z Waiting for 1s while stateStr !== PRIMARY, current: SECONDARY
mongo-rs-init  | 2025-06-10T17:56:54.442933009Z {
mongo-rs-init  | 2025-06-10T17:56:54.442947549Z   set: 'rs0',
mongo-rs-init  | 2025-06-10T17:56:54.442952309Z   date: ISODate('2025-06-10T17:56:54.431Z'),
mongo-rs-init  | 2025-06-10T17:56:54.442955979Z   myState: 1,
mongo-rs-init  | 2025-06-10T17:56:54.442959419Z   term: Long('1'),
mongo-rs-init  | 2025-06-10T17:56:54.442962789Z   syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:54.442965949Z   syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:54.442969540Z   heartbeatIntervalMillis: Long('2000'),
mongo-rs-init  | 2025-06-10T17:56:54.442973720Z   majorityVoteCount: 2,
mongo-rs-init  | 2025-06-10T17:56:54.442977910Z   writeMajorityCount: 2,
mongo-rs-init  | 2025-06-10T17:56:54.442981960Z   votingMembersCount: 3,
mongo-rs-init  | 2025-06-10T17:56:54.442984940Z   writableVotingMembersCount: 3,
mongo-rs-init  | 2025-06-10T17:56:54.442987780Z   optimes: {
mongo-rs-init  | 2025-06-10T17:56:54.442990690Z     lastCommittedOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.442993690Z     lastCommittedWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:54.442996590Z     readConcernMajorityOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.442999770Z     appliedOpTime: { ts: Timestamp({ t: 1749578214, i: 2 }), t: Long('1') },
mongo-rs-init  | 2025-06-10T17:56:54.443002840Z     durableOpTime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.443005750Z     lastAppliedWallTime: ISODate('2025-06-10T17:56:54.386Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443008650Z     lastDurableWallTime: ISODate('2025-06-10T17:56:42.828Z')
mongo-rs-init  | 2025-06-10T17:56:54.443011540Z   },
mongo-rs-init  | 2025-06-10T17:56:54.443014390Z   lastStableRecoveryTimestamp: Timestamp({ t: 1749578202, i: 1 }),
mongo-rs-init  | 2025-06-10T17:56:54.443017350Z   electionCandidateMetrics: {
mongo-rs-init  | 2025-06-10T17:56:54.443020240Z     lastElectionReason: 'electionTimeout',
mongo-rs-init  | 2025-06-10T17:56:54.443023090Z     lastElectionDate: ISODate('2025-06-10T17:56:54.134Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443025980Z     electionTerm: Long('1'),
mongo-rs-init  | 2025-06-10T17:56:54.443028830Z     lastCommittedOpTimeAtElection: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.443032821Z     lastSeenOpTimeAtElection: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.443035751Z     numVotesNeeded: 2,
mongo-rs-init  | 2025-06-10T17:56:54.443038531Z     priorityAtElection: 30,
mongo-rs-init  | 2025-06-10T17:56:54.443041291Z     electionTimeoutMillis: Long('10000'),
mongo-rs-init  | 2025-06-10T17:56:54.443044101Z     numCatchUpOps: Long('0')
mongo-rs-init  | 2025-06-10T17:56:54.443046861Z   },
mongo-rs-init  | 2025-06-10T17:56:54.443049541Z   members: [
mongo-rs-init  | 2025-06-10T17:56:54.443052321Z     {
mongo-rs-init  | 2025-06-10T17:56:54.443055121Z       _id: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443057811Z       name: 'mongo1:27017',
mongo-rs-init  | 2025-06-10T17:56:54.443060571Z       health: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443067871Z       state: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443071021Z       stateStr: 'PRIMARY',
mongo-rs-init  | 2025-06-10T17:56:54.443073981Z       uptime: 17,
mongo-rs-init  | 2025-06-10T17:56:54.443076881Z       optime: { ts: Timestamp({ t: 1749578214, i: 2 }), t: Long('1') },
mongo-rs-init  | 2025-06-10T17:56:54.443079831Z       optimeDate: ISODate('2025-06-10T17:56:54.000Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443082721Z       lastAppliedWallTime: ISODate('2025-06-10T17:56:54.386Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443085631Z       lastDurableWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443088501Z       syncSourceHost: '',
mongo-rs-init  | 2025-06-10T17:56:54.443091331Z       syncSourceId: -1,
mongo-rs-init  | 2025-06-10T17:56:54.443094161Z       infoMessage: 'Could not find member to sync from',
mongo-rs-init  | 2025-06-10T17:56:54.443097041Z       electionTime: Timestamp({ t: 1749578214, i: 1 }),
mongo-rs-init  | 2025-06-10T17:56:54.443100312Z       electionDate: ISODate('2025-06-10T17:56:54.000Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443103172Z       configVersion: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443105952Z       configTerm: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443108742Z       self: true,
mongo-rs-init  | 2025-06-10T17:56:54.443112262Z       lastHeartbeatMessage: ''
mongo-rs-init  | 2025-06-10T17:56:54.443115162Z     },
mongo-rs-init  | 2025-06-10T17:56:54.443117982Z     {
mongo-rs-init  | 2025-06-10T17:56:54.443120752Z       _id: 2,
mongo-rs-init  | 2025-06-10T17:56:54.443123562Z       name: 'mongo2:27018',
mongo-rs-init  | 2025-06-10T17:56:54.443126382Z       health: 1,
mongo-rs-init  | 2025-06-10T17:56:54.443129262Z       state: 2,
mongo-rs-init  | 2025-06-10T17:56:54.443132042Z       stateStr: 'SECONDARY',
mongo-rs-init  | 2025-06-10T17:56:54.443134862Z       uptime: 11,
mongo-rs-init  | 2025-06-10T17:56:54.443137742Z       optime: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.443140682Z       optimeDurable: { ts: Timestamp({ t: 1749578202, i: 1 }), t: Long('-1') },
mongo-rs-init  | 2025-06-10T17:56:54.443143592Z       optimeDate: ISODate('2025-06-10T17:56:42.000Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443146602Z       optimeDurableDate: ISODate('2025-06-10T17:56:42.000Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443150632Z       lastAppliedWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443154452Z       lastDurableWallTime: ISODate('2025-06-10T17:56:42.828Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443157572Z       lastHeartbeat: ISODate('2025-06-10T17:56:54.278Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443160722Z       lastHeartbeatRecv: ISODate('2025-06-10T17:56:54.277Z'),
mongo-rs-init  | 2025-06-10T17:56:54.443163873Z       pingMs: Long('0'),
mongo-rs-init  | 2025-06-10T17:56:54.443166953Z       lastHeartbeatMessage: '',
…